fix(easyflash): 完成移植，并修复eeprom空间配置
eeprom_onRTC：修复空间错误，非256K 实际空间32k

+++ bsp/lpc408x/drivers/FM31XX/src/fm31XX.c
@@ -19,8 +19,8 @@
 // Fm31xx_port.c
 void fm31xx_hw_i2c_init();
 
-int Fm31xxWriteReg(uint8_t reg, uint8_t *pdt, uint8_t len);
-int Fm31xxReadReg(uint8_t reg, uint8_t *pdt, uint8_t len);
+int Fm31xxWriteReg(uint8_t reg, uint8_t *pdt, uint32_t len);
+int Fm31xxReadReg(uint8_t reg, uint8_t *pdt, uint32_t len);
 /**********************************************************************************
 ** 函数名称:
 ** 功    能:
